backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
groq==0.4.1
requests==2.31.0
beautifulsoup4==4.12.2
jinja2==3.1.2
python-multipart==0.0.6
httpx==0.25.2
lxml==4.9.3
aiofiles==23.2.1
pillow>=10.4.0
python-dotenv==1.0.0
playwright==1.58.0
